refactor(cartesiandomain): move doubly periodic boundary conditions into seperate namespace
