+++ libs/cartesiandomain/cartesianmotion.hpp
@@ -31,8 +31,8 @@
 #include <cassert>
 #include <functional>
 
-#include "cartesiandomain/cartesiandomainboundaries.hpp"
 #include "cartesiandomain/cartesianmaps.hpp"
+#include "cartesiandomain/doubly_periodic_domain.hpp"
 #include "gridboxes/predcorrmotion.hpp"
 #include "superdrops/superdrop.hpp"
 #include "superdrops/terminalvelocity.hpp"
