+++ libs/cartesiandomain/cartesianmotion_withreset.hpp
@@ -36,7 +36,6 @@
 
 #include "../cleoconstants.hpp"
 #include "../kokkosaliases.hpp"
-#include "cartesiandomain/cartesiandomainboundaries.hpp"
 #include "cartesiandomain/cartesianmaps.hpp"
 #include "cartesiandomain/cartesianmotion.hpp"
 #include "gridboxes/predcorrmotion.hpp"
